[DevTools] chore: extract getInternalReactConstants
